# docs/swagger.yaml
definitions:
...
  dto.AdminLoginInput:
    properties:
      password:
        description: password
        example: "123456"
        type: string
      username:
        description: username
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.AdminLoginOutput:
    properties:
      token:
        description: token
        example: token
        type: string
    type: object
  dto.ChangePwdInput:
    properties:
      password:
        description: password
        example: "123456"
        type: string
    required:
    - password
    type: object
...
  dto.ServiceStatOutput:
    properties:
      today:
        description: list
        items:
          type: integer
        type: array
      yesterday:
        description: list
        items:
          type: integer
        type: array
    type: object
...
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: Admin Change Password
      operationId: /admin/change_pwd
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePwdInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Admin Change Password
      tags:
      - Admin interface
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: Admin Login
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminLoginOutput'
              type: object
      summary: Admin Login
      tags:
      - Admin interface
...
  /service/service_stat:
    get:
      consumes:
      - application/json
      description: ServiceStat
      operationId: /service/service_stat
      parameters:
      - description: service ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceStatOutput'
              type: object
      summary: ServiceStat
      tags:
      - Service Interface
...
swagger: "2.0"